+++ FBSDKCoreKit/FBSDKCoreKitTests/GraphRequestTests.swift
@@ -24,7 +24,7 @@ final class GraphRequestTests: XCTestCase {
 
   let path = "me"
   let parameters = ["fields": ""]
-  let version = "v11.0"
+  let version = "v12.0"
   let prefix = "graph."
   let settings = TestSettings()
   var factory = TestGraphRequestConnectionFactory()